# backend/jvm/webapi/src/main/resources/openapi.yaml
openapi: 3.0.0
info:
  title: Ticket Management API
  version: 1.0.0
servers:
  - url: http://localhost:8080/
...
paths:
...
  /signup:
    post:
      tags:
        - Authentication
      summary: Create a new user
      operationId: createUser
      requestBody:
        description: User object to be created
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserPost'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
...
components:
...
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
  schemas:
...
    User:
      type: object
      properties:
        userId:
          type: string
          format: uuid
          readOnly: true
        username:
          type: string
        fname:
          type: string
        lname:
          type: string
        email:
          type: string
        password:
          type: string
    UserPost:
      type: object
      properties:
        username:
          type: string
        fname:
          type: string
        lname:
          type: string
        email:
          type: string
        password:
          type: string
        role:
          type: string
          enum:
            - USER
            - EMPLOYEE
            - ADMIN
          default: USER
...